requirements.txt
torch
torchvision
optuna
PyYAML
mlflow